+++ vitest.config.ts
@@ -7,6 +7,14 @@ export default defineConfig({
     include: ['src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
     coverage: {
       provider: 'v8',
+      include: ['src/**/*.ts'],
+      exclude: [
+        'src/tests/**',
+        'src/**/*.test.ts',
+        'src/**/*.spec.ts',
+        'src/index.ts',
+        'debugging/**',
+      ],
       reporter: ['text', 'json', 'html'],
     },
   },